eloquent

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -20,8 +20,8 @@ return new class extends Migration
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password', 255);
             $table->string('hak_akses', 25);
-            $table->integer('unit_kerja_id');
-            $table->integer('jabatan_id');
+            $table->bigInteger('unit_kerja_id');
+            $table->bigInteger('jabatan_id');
             $table->rememberToken();
             $table->timestamps();
         });
